Pushing the initial project with basic functionality
Started working on a little p5js-powered tool called DeMa - pronounced like the short form of Dmitry - Design Materializer. This tool is intended to support grounded theory analysis of git repositories.
Basic functionality for displaying commits along a timeline is there. Currently working on the coding feature, i.e., the ability to add codes to commits.
I will also need to add a saving function that will store all the commitobjs. When that is implemented, the user should be able to select between opening a new JSON commit dump or loading a save file.
Once that works, I want to add support for displaying markdown journals (including illustrations!), based on an entries date. The user should be able to code these as well.

Also uploaded a small python script that must be run on a cloned repository, to get the initial JSON commit dump file for import into DeMa.